+++ src/OpenTelemetry.Exporter.Prometheus.HttpListener/Internal/PrometheusCollectionManager.cs
@@ -173,6 +173,8 @@ internal sealed class PrometheusCollectionManager
 
         try
         {
+            var resourceTags = new PrometheusResourceTagCollection(this.exporter.Resource, this.exporter.AllowedResourceAttributesFilter);
+
             if (this.exporter.OpenMetricsRequested)
             {
                 cursor = this.WriteTargetInfo();
@@ -230,6 +232,7 @@ internal sealed class PrometheusCollectionManager
                             cursor,
                             metric,
                             this.GetPrometheusMetric(metric),
+                            resourceTags,
                             this.exporter.OpenMetricsRequested);
 
                         break;