+++ src/OpenTelemetry.Exporter.Prometheus.HttpListener/Internal/PrometheusExporterOptions.cs
@@ -33,4 +33,9 @@ internal sealed class PrometheusExporterOptions
     /// Gets or sets a value indicating whether addition of _total suffix for counter metric names is disabled. Default value: <see langword="false"/>.
     /// </summary>
     public bool DisableTotalNameSuffixForCounters { get; set; }
+
+    /// <summary>
+    /// Gets or sets the allowed resource attributes filter. Default value: null (no attributes allowed).
+    /// </summary>
+    public Predicate<string> AllowedResourceAttributesFilter { get; set; } = null;
 }