+++ test/OpenTelemetry.Exporter.Prometheus.HttpListener.Tests/PrometheusHttpListenerTests.cs
@@ -85,6 +85,12 @@ public class PrometheusHttpListenerTests
     }
 
     [Fact]
+    public async Task PrometheusExporterHttpServerIntegration_AddResourceAttributeAsTag()
+    {
+        await this.RunPrometheusExporterHttpServerIntegrationTest(addServiceNameResourceTag: true);
+    }
+
+    [Fact]
     public void PrometheusHttpListenerThrowsOnStart()
     {
         Random random = new Random();
@@ -155,7 +161,7 @@ public class PrometheusHttpListenerTests
             });
     }
 
-    private async Task RunPrometheusExporterHttpServerIntegrationTest(bool skipMetrics = false, string acceptHeader = "application/openmetrics-text")
+    private async Task RunPrometheusExporterHttpServerIntegrationTest(bool skipMetrics = false, string acceptHeader = "application/openmetrics-text", bool addServiceNameResourceTag = false)
     {
         var requestOpenMetrics = acceptHeader.StartsWith("application/openmetrics-text");
 
@@ -180,6 +186,11 @@ public class PrometheusHttpListenerTests
                     .AddPrometheusHttpListener(options =>
                     {
                         options.UriPrefixes = new string[] { address };
+
+                        if (addServiceNameResourceTag)
+                        {
+                            options.AllowedResourceAttributesFilter = s => s == "service.name";
+                        }
                     })
                     .Build();
 
@@ -234,6 +245,10 @@ public class PrometheusHttpListenerTests
 
             var content = await response.Content.ReadAsStringAsync();
 
+            var resourceTagAttributes = addServiceNameResourceTag
+                ? "service_name='my_service',"
+                : string.Empty;
+
             var expected = requestOpenMetrics
                 ? "# TYPE target info\n"
                   + "# HELP target Target metadata\n"
@@ -242,10 +257,10 @@ public class PrometheusHttpListenerTests
                   + "# HELP otel_scope_info Scope metadata\n"
                   + $"otel_scope_info{{otel_scope_name='{MeterName}'}} 1\n"
                   + "# TYPE counter_double_total counter\n"
-                  + $"counter_double_total{{otel_scope_name='{MeterName}',otel_scope_version='{MeterVersion}',key1='value1',key2='value2'}} 101.17 (\\d+\\.\\d{{3}})\n"
+                  + $"counter_double_total{{{resourceTagAttributes}otel_scope_name='{MeterName}',otel_scope_version='{MeterVersion}',key1='value1',key2='value2'}} 101.17 (\\d+\\.\\d{{3}})\n"
                   + "# EOF\n"
                 : "# TYPE counter_double_total counter\n"
-                  + $"counter_double_total{{otel_scope_name='{MeterName}',otel_scope_version='{MeterVersion}',key1='value1',key2='value2'}} 101.17 (\\d+)\n"
+                  + $"counter_double_total{{{resourceTagAttributes}otel_scope_name='{MeterName}',otel_scope_version='{MeterVersion}',key1='value1',key2='value2'}} 101.17 (\\d+)\n"
                   + "# EOF\n";
 
             Assert.Matches(("^" + expected + "$").Replace('\'', '"'), content);