+++ src/OpenTelemetry.Exporter.Prometheus.HttpListener/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Add resource attributes as tags for Prometheus exporters with filter
+  ([#3087](https://github.com/open-telemetry/opentelemetry-dotnet/pull/5407))
+
 ## 1.8.0-rc.1
 
 Released 2024-Mar-27